+++ mq/mq_test.go
@@ -12,9 +12,9 @@ import (
 	"time"
 	"unsafe"
 
-	"bitbucket.org/avd/go-ipc"
 	"bitbucket.org/avd/go-ipc/internal/allocator"
-	"bitbucket.org/avd/go-ipc/internal/test"
+	"bitbucket.org/avd/go-ipc/internal/common"
+	testutil "bitbucket.org/avd/go-ipc/internal/test"
 	ipc_sync "bitbucket.org/avd/go-ipc/sync"
 
 	"github.com/stretchr/testify/assert"
@@ -95,7 +95,7 @@ func testCreateMqExcl(t *testing.T, ctor mqCtor, dtor mqDtor) {
 	}
 	_, err = ctor(testMqName, os.O_EXCL, 0666)
 	a.Error(err)
-	if d, ok := mq.(ipc.Destroyer); ok {
+	if d, ok := mq.(common.Destroyer); ok {
 		a.NoError(d.Destroy())
 	} else {
 		a.NoError(mq.Close())